Add a placeholder special page
`Special:DoesUserExist` will add a way to check if a user exists with a given
name. For now it just outputs an HTML message to demonstrate setting up a new
special page.

Additionally, add an extension messages file to set up aliases - needed for the
special page, as well as other features in the future.