SpecialDoesUserExist: add handling for hidden users
If a user with the requested name exists but is hidden, for viewers without the
`hideuser` right just pretend like the user does not exist. If the user is
hidden but the viewer *does* have the `hideuser` right, confirm the user's
existence but also add a note that they are hidden.