Add a placeholder API module
The `doesuserexist` API module will work similarly to the special page as a
simple function that demonstrates how extensions can add to the MediaWiki
interface. For now, it just reports that it is not implemented yet, serving to
separate demonstration of how to register API modules from demonstration of the
actual implementation.

In the process, add another i18n directory for the API modules specifically in
keeping with the practice of other extensions.