ApiDoesUserExist: add examples
Add examples of a user existing, existing but with the requested name needing
to be normalized, and not existing. These examples assume that the user
"MediaWiki default" will always exist (since it is used by the installer) and
that the user "Foo/bar" will never exist (since it has a slash and should be
treated as invalid).