ParserHooks: add initial implementation of PAGECONTENTMODEL
The parser function accepts a single parameter, the name of a page to get the
content model of. If no name is provided, the current page is assumed instead,
and is assumed to be wikitext. If the page name provided is invalid, or for a
page that does not exist, an empty string is returned. Otherwise, we get the
actual content model and return it.