Add placeholder markdown content
Currently just exists as a wrapper around text content, has no actual extra
rendering or other logic. This is just to demonstrate registration of a new
content model, the actual implementation will come in a follow-up commit.

Also includes the i18n message to localize the content model name - currently
for English it will show as "Markdown", e.g. as the option name on
Special:ChangeContentModel.

The classes are in a dedicated sub-directory (and thus in a dedicated
sub-namespace) since there will be a number of classes added to support this
single feature.