Markdown: add styles for external links
Switch from just using the `CommonMarkConverter` to building our own
`Environment` that includes both the `CommonMarkCoreExtension` and another
extension, `ExternalLinkExtension`, which allows manipulating external links.
Use that environment to first parse the raw markdown text, and then render it.
Configure the added `ExternalLinkExtension` to apply the `external` CSS class,
and wrap the overall parser output with the `mw-parser-output` class so that
the core styles for external links are applied.