Markdown: add image handling
For images where the source is given as a URL, just stop the image from being
loaded. We could add support for MediaWiki core's various exceptions
(configuration options `AllowExternalImages`, `AllowExternalImagesFrom`,
and `EnableImageWhitelist`) but that seems unneeded at this point since we
just want to demonstrate how to have the content model interact with MediaWiki
core.

We also make sure to inform MediaWiki that the images are used on a given page,
just as we did for links. However, since we are using MediaWiki's own parser
for rendering the images, telling MediaWiki about their uses is much easier.