Add an option to hide the note on Special:ChangeContentModel
Serves as an example of registering user options. This may not be the most
useful preference to add for a real extension, since content models are
generally rarely changed and the note is not very big, and every new preference
adds complexity. See the essay [[Just make it a user preference]] on
MediaWiki.org for a good explanation of the problems new preferences can pose.
But, it is a reasonable preference for this example extension and a good
demonstration of how preferences are added and used.