Add a placeholder table for syntax help
Create a new database table, `me_syntaxhelp`, that will associate content
models with pages for getting help with the syntax of that content model.
At the moment, entirely unused, just a demonstration of adding the table.

The classes are in a dedicated sub-directory (and thus in a dedicated
sub-namespace) since there will be a number of classes added to support this
single feature.