SyntaxHelp: add API endpoint to identify page
Add a new API action, `getsyntaxhelp`, to retrieve the name of the page that
provides help with syntax for a specific content model. This provides a way
other than the configuration page itself to indentify where to get help for
syntax, and will be connected to a ResourceLoader module in a subsequent
commit.