SyntaxHelp: add ResourceLoader module to add help link
When a page is being edited (action is 'edit' or 'submit'), add the
`ext.minimalexample.syntaxhelp` ResourceLoader module. That module will query
the `getsyntaxhelp` API endpoint to determine if there is a page which provides
help for the syntax of the current content model; if there is such a page, a
link to it is added after the page title.

The link is currently put directly after the page title, and is shown with the
same styles as the title itself. In a follow-up commit I will add some CSS to
add spacing between the title and the link, and style the link.