SyntaxHelp: restrict who can configure the help pages
Add a new user right, `syntaxhelp-configure`, that by default is only granted
to administrators. Require this right to be able to access
SpecialConfigureSyntaxHelp at all.

This is a coarse-grained approach that also prevents users without the right
from viewing the page to see all of the associated help pages at the same time.
A follow-up commit will change it so that the right is only needed to actually
change the configuration, and users without the right will be able to view the
configuration.