Tests: add unit tests for ParserHooks
Unit tests use mocking and are helpful to ensure that all of the code paths can
be executed properly, but are not the most useful since in some cases they
essentially duplicate the implementation logic. A follow-up commit will add
integration tests to ensure that in a production-like environment the code
still works as expected.