Create ApiDoesUserExistTest.php
Tests: add integration tests for ApiDoesUserExist

Test that when checking for a specific user, the output is correct regarding if
the user exists, the name gets normalized, the user is hidden, etc. Also
confirm that multiple users can be checked at once, and that warnings are added
for duplicate requests.