ExtraDetails: implement new feature
For users with the `edit-extra-details` right (granted to administrators by
default), a new API module, accessed via action=setextradetails, can be used to
set the contents of a page slot for storing some extra details. That content is
parsed as wikitext and shown at the bottom of the page to logged in users;
anonymous users can force it to show up with the `force-show-extra-details`
query parameter.

There are not yet tests for the API module or rendering logic.